Move compiler related source code into `compiler` package.
