+++ casper/src/test/scala/coop/rchain/casper/genesis/contracts/TestUtil.scala
@@ -7,7 +7,8 @@ import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.models.Par
 import coop.rchain.rholang.build.CompiledRholangSource
 import coop.rchain.rholang.interpreter.accounting.Cost
-import coop.rchain.rholang.interpreter.{ParBuilder, Runtime}
+import coop.rchain.rholang.interpreter.Runtime
+import coop.rchain.rholang.interpreter.compiler.ParBuilder
 
 object TestUtil {
   def eval[F[_]: Sync, Env](
