+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/ReceiveBindsSortMatcher.scala
@@ -1,8 +1,8 @@
-package coop.rchain.rholang.interpreter
+package coop.rchain.rholang.interpreter.compiler
 
 import cats.effect.Sync
-import coop.rchain.models.{Par, ReceiveBind, Var}
 import coop.rchain.models.rholang.sorter.ReceiveSortMatcher.sortBind
+import coop.rchain.models.{Par, ReceiveBind, Var}
 import cats.implicits._
 import coop.rchain.models.rholang.sorter._
 import coop.rchain.models.rholang.implicits._
